pyproject.toml
[project]
name = "dobble-gen"
version = "0.1.0"
description = "Generate your own Dobble cards! Ready to be printed!"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=12.0.0",
    "random2>=1.0.2",
    "reportlab>=4.4.4",
    "typer>=0.20.0",
]

[project.scripts]
dobble-gen = "dobble_gen:cli.main"

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.18.2",
    "pytest>=9.0.1",
    "ruff>=0.14.4",
    "types-reportlab>=4.4.4.20250926",
]
